Health Bar Manager And Behavior
Configured HealthBarManager and created HealthBarBehavior. The Manager handles the creation of the health bar and assigning initial and essential variables that the health bar will use. The behavior script has been given many methods to assign values, ensuring they're not the wrong type of data. It gets a reference to the enemy and data to be able to follow the enemy while also keeping track of the health, updating the slider accordingly.